compile zisk
